requirements.txt
RPi.GPIO
w1thermsensor
adafruit-circuitpython-bme280
paho-mqtt
Pillow
neopixel
